really remove the navigation object

+++ src/main/com/topcoder/web/tc/controller/legacy/pacts/bean/pacts_client/dispatch/PaymentBean.java
@@ -54,7 +54,7 @@ public class PaymentBean implements PactsConstants {
      * @param memberId the member id
      * @return the payment headers, or null if the payment id is bad
      */
-    public PaymentHeader[] getPaymentsForUser(int memberId) {
+    public PaymentHeader[] getPaymentsForUser(long memberId) {
         log.debug("getPaymentsForUser, memberId = " + memberId);
         DataInterfaceBean bean = new DataInterfaceBean();
         java.util.Map reply = null;
